refactor(navbar): mobile menu component
