+++ components/navbar/MobileMenu.js
@@ -1,0 +1,60 @@
+import * as React from "react";
+import {Box, IconButton, Typography, Menu, MenuItem} from "@mui/material";
+import MenuIcon from "@mui/icons-material/Menu";
+
+function MobileMenu({pages}) {
+  const [anchorElNav, setAnchorElNav] = React.useState(null);
+  
+  const handleOpenNavMenu = (event) => {
+    setAnchorElNav(event.currentTarget);
+  };
+  
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+  
+  return(
+    <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
+      <IconButton
+        size="large"
+        aria-label="account of current user"
+        aria-controls="menu-appbar"
+        aria-haspopup="true"
+        onClick={handleOpenNavMenu}
+        color="inherit"
+      >
+        <MenuIcon />
+      </IconButton>
+      <Menu
+        id="menu-appbar"
+        anchorEl={anchorElNav}
+        anchorOrigin={{
+          vertical: "bottom",
+          horizontal: "left",
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: "top",
+          horizontal: "left",
+        }}
+        open={Boolean(anchorElNav)}
+        onClose={handleCloseNavMenu}
+        sx={{
+          display: { xs: "block", md: "none" },
+        }}
+      >
+        {pages.map((page) => (
+          <MenuItem
+            key={page.text}
+            onClick={handleCloseNavMenu}
+            href={page.url}
+          >
+            <Typography textAlign="center">{page.text}</Typography>
+          </MenuItem>
+        ))}
+      </Menu>
+    </Box>
+  );
+}
+
+export default MobileMenu;
