+++ components/navbar/Navbar.js
@@ -1,28 +1,19 @@
 import * as React from "react";
-import {AppBar, CircularProgress, Box, Toolbar, IconButton, Typography, Menu, Container, Avatar, Button, Tooltip, MenuItem} from "@mui/material";
-import MenuIcon from "@mui/icons-material/Menu";
-import { useOperation, useOperationMethod } from "react-openapi-client";
+import {AppBar, CircularProgress, Toolbar, Container} from "@mui/material";
+import { useOperation } from "react-openapi-client";
 import AppTitleAndLogo from "./AppTitleAndLogo";
 import UserBubble from "./UserBubble";
 import DesktopMenu from "./DesktopMenu";
+import MobileMenu from "./MobileMenu";
 
 const pages = [{ text: "Home", url: "/" }];
 
 function ResponsiveAppBar() {
-  const [anchorElNav, setAnchorElNav] = React.useState(null);
   const { loading, data, error } = useOperation("getSession");
 
   let userName = null;
   let userId = 0;
-  let settings = {};
-
-  const handleOpenNavMenu = (event) => {
-    setAnchorElNav(event.currentTarget);
-  };
-
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
+  let sessionActions = {};
 
   if (loading) {
     return <div><CircularProgress /></div>;
@@ -31,14 +22,14 @@ function ResponsiveAppBar() {
   if (error) {
     userName = null;
     userId = 0;
-    settings = [
+    sessionActions = [
       { text: "Login", url: "users/sign_in" },
       { text: "Signup", url: "users/sign_up" },
     ];
   } else {
     userName = data.name;
     userId = data.id;
-    settings = [{ text: "Logout", url: null }];
+    sessionActions = [{ text: "Logout", url: null }];
   }
 
   return (
@@ -46,50 +37,10 @@ function ResponsiveAppBar() {
       <Container maxWidth="xl">
         <Toolbar disableGutters>
           <AppTitleAndLogo display={{ xs: "none", md: "flex" }} variant="h6" flexGrow={0} />
-
-          <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              color="inherit"
-            >
-              <MenuIcon />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "left",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "left",
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: "block", md: "none" },
-              }}
-            >
-              {pages.map((page) => (
-                <MenuItem
-                  key={page.text}
-                  onClick={handleCloseNavMenu}
-                  href={page.url}
-                >
-                  <Typography textAlign="center">{page.text}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
+          <MobileMenu pages={pages}/>
           <AppTitleAndLogo display={{ xs: "flex", md: "none" }} variant ="h5" flexGrow={1}/>
           <DesktopMenu pages={pages}/>
-          <UserBubble sessionActions={settings} userName={userName} userId={userId}/>
+          <UserBubble sessionActions={sessionActions} userName={userName} userId={userId}/>
         </Toolbar>
       </Container>
     </AppBar>
